requirements.txt
# Selenium for browser automation
selenium==4.16.0

# Image processing with EXIF support
Pillow==10.1.0

# Automatic ChromeDriver management
webdriver-manager==4.0.1

# Environment variable management
python-dotenv==1.0.0

# Rich console output and progress bars
rich==13.7.0
